requirements.txt
Appium-Python-Client==2.7.1
furl==2.1.3
h2==4.1.0
h11==0.14.0
killport==1.2.0
mattermostdriver==7.3.2
mitmproxy==9.0.1
portforward==0.4.5
psycopg2-binary==2.9.6
pytest==7.2.0
pytest-check==1.3.0
pytest-html==3.2.0
pytest-timeout==2.1.0
requests==2.28.1
selenium==4.7.2
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
exceptiongroup==1.0.4
Flask==2.2.2
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.1
mitmproxy_wireguard==0.1.19
msgpack==1.0.4
orderedmultidict==1.0.1
outcome==1.2.0
packaging==22.0
passlib==1.7.4
pip==21.3.1
pluggy==1.0.0
protobuf==4.21.12
psutil==5.9.4
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pymongo==3.10.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
pyTelegramBotAPI==4.8.0
pytest-metadata==2.0.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
setuptools==60.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
urwid==2.1.2
websockets==10.4
Werkzeug==2.2.2
wheel==0.37.1
wsproto==1.2.0
zstandard==0.19.0